Refactor `card`s used in `CardGrid`
