+++ src/stories/Library/card-grid/card-grid-data.tsx
@@ -1,0 +1,25 @@
+import Card from "../card/Card";
+import ImageCredited from "../image-credited/ImageCredited";
+
+const imageUrl =
+  "https://images.unsplash.com/photo-1568667256549-094345857637?q=80&w=2815&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D";
+
+const image = <ImageCredited src={imageUrl} />;
+
+export const card = (
+  <Card
+    title="Bøger som har gjort en forskel for romanens udvikling"
+    typeTag="Nyhed"
+    dateTag="06 Dec 2022"
+    image={image}
+  />
+);
+
+export const cardNoImage = (
+  <Card
+    title="Fars legestue, hver onsdag"
+    typeTag="Arrangement"
+    dateTag="06 Okt - 28 Dec 2022"
+    placeholderText="Fri leg for alle aldre"
+  />
+);
